+++ ChangeLog.md
@@ -36,6 +36,8 @@
 - ADDED: #4417 Html Export plugin produces now HTML tables that are sortable by clicking on their headers.
 - ADDED: #4286 SQL History tab has now function to export executed queries history (or selected subset) to a SQL file.
 - ADDED: #5369 The Format SQL action added to context menus everywhere where SQL is being being edited. Also the hotkey for actions works for it.
+- ADDED: #4401 Custom footer and header can be configured for PDF and HTML export.
+- ADDED: #4401 PDF export now includes query used for the data being exported (when exporting Query Results or a View).
 - CHANGE: #4553 Application has moved to Qt 6 (dropping Qt 5 support). This solves several technical issues (for example accented characters on MacOS, better UI scaling for high DPI displays, better dark theme support and more).
 - CHANGE: #5102 All raster icons are replaced with vector icons (SVG). This makes them scalable and work well with high DPI displays.
 - CHANGE: #4271 Value Editor (and Form View) got small visual lifting to be more convenient to use (bigger column labels, etc).